Extract an AccountSwitcher component
